refactor: simplify response logic for asynchronous elements into multiple functions
